Improve ELF offset calculation
Previously, the offset of a virtual address in the ELF file was
calculated as `v_addr - v_load_addr`, where `v_load_addr` is the virtual
address where the dynamic object is loaded, as returned by dladdr. This
calculation incorrectly assumes that the "base offset" in the ELF is 0,
which is not always true. This caused `addr2line` to produce bogus
output on fedora.

The new calculation relies on `dladdr1`, which can return a `link_map`
object that contains the difference between the virtual address and the
location in the ELF file.

A minor index out of bounds issue was also fixed, and debug logging was
added.